+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -67,6 +67,8 @@ import java.util.logging.Logger;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManagerFactory;
 
+import io.netty.handler.codec.http.cors.CorsConfig;
+
 import com.vmware.xenon.common.FileUtils.ResourceEntry;
 import com.vmware.xenon.common.NodeSelectorService.SelectAndForwardRequest;
 import com.vmware.xenon.common.NodeSelectorService.SelectAndForwardRequest.ForwardingOption;
@@ -1449,6 +1451,7 @@ public class ServiceHost implements ServiceRequestSender {
                 this.httpListener.setResponsePayloadSizeLimit(this.state.responsePayloadSizeLimit);
             }
 
+            this.httpListener.setCorsConfig(getCorsConfig());
             this.httpListener.start(getPort(), this.state.bindAddress);
         }
 
@@ -1472,6 +1475,7 @@ public class ServiceHost implements ServiceRequestSender {
                             .setResponsePayloadSizeLimit(this.state.responsePayloadSizeLimit);
                 }
 
+                this.httpsListener.setCorsConfig(getCorsConfig());
                 this.httpsListener.start(getSecurePort(), this.state.bindAddress);
             }
         }
@@ -6288,4 +6292,29 @@ public class ServiceHost implements ServiceRequestSender {
                 ServiceErrorResponse.ERROR_CODE_SERVICE_ALREADY_EXISTS,
                 e);
     }
+
+    /**
+     * CORS configuration for netty pipeline.
+     *
+     * By default, CORS is disabled.
+     * To enable CORS support, override this method in subclass and populate {@link CorsConfig}.
+     *
+     * Example:
+     * {@code
+     *   // allow PUT for any origin
+     *   CorsConfigBuilder.forAnyOrigin().allowedRequestMethods(HttpMethod.PUT).build();
+     * }
+     *
+     * When {@code null} is returned, CORS support from nettty(by {@link io.netty.handler.codec.http.cors.CorsHandler})
+     * will be disabled.
+     *
+     * @return CORS configuration
+     * @see io.netty.handler.codec.http.cors.CorsConfigBuilder
+     * @see io.netty.handler.codec.http.cors.CorsConfig
+     */
+    public CorsConfig getCorsConfig() {
+        // CORS is disabled by default
+        return null;
+    }
+
 }